[REFACTOR] Moved file to same package as other log files
